+++ web/app/loader.js
@@ -29,6 +29,19 @@ const loadWorker = () => (
         .catch(() => {})
 )
 
+const loadMain = (target) => {
+    const script = document.createElement('script')
+    script.id = 'progressive-web-script'
+    // Setting UTF-8 as our encoding ensures that certain strings (i.e.
+    // Japanese text) are not improperly converted to something else.
+    script.charset = 'utf-8'
+    script.src = getAssetUrl('main.js')
+
+    if (target) {
+        target.appendChild(script)
+    }
+}
+
 if (isReactRoute()) {
     displayPreloader(preloadCSS, preloadHTML, preloadJS)
 
@@ -71,13 +84,30 @@ if (isReactRoute()) {
             rel: 'manifest'
         })
 
-        const script = document.createElement('script')
-        script.id = 'progressive-web-script'
+        // This method is defined so that when main.js loads, it doesn't attempt
+        // to load its dependencies synchronously, because it's dependencies
+        // in vendor.js may have not loaded yet. We replace the call to
+        // webpackJsonp in main.js with this webpackJsonpAsync function, which
+        // will wait for webpackJsonp (and thus, vendor.js) to be finished loading.
+        window.webpackJsonpAsync = (module, exports, webpackRequire) => {
+            const runJsonpAsync = function() {
+                return window.webpackJsonp ?
+                    window.webpackJsonp(module, exports, webpackRequire) :
+                    setTimeout(runJsonpAsync, 50)
+            }
+
+            runJsonpAsync()
+        }
+
+        const vendorScript = document.createElement('script')
+        vendorScript.id = 'progressive-web-script'
         // Setting UTF-8 as our encoding ensures that certain strings (i.e.
-        // Japanese text) are not improperly converted to something else.
-        script.charset = 'utf-8'
-        script.src = getAssetUrl('main.js')
-        body.appendChild(script)
+        // Japanese text) are not improperly converted to something else. We
+        // do this on the vendor scripts also just in case any libs we import
+        // have localized strings in them.
+        vendorScript.charset = 'utf-8'
+        vendorScript.src = getAssetUrl('vendor.js')
+        body.appendChild(vendorScript)
 
         const jQuery = document.createElement('script')
         jQuery.async = true
@@ -85,11 +115,23 @@ if (isReactRoute()) {
         jQuery.src = getAssetUrl('static/js/jquery.min.js')
         body.appendChild(jQuery)
 
-        const capturing = document.createElement('script')
-        capturing.async = true
-        capturing.id = 'progressive-web-capture'
-        capturing.src = CAPTURING_CDN
-        body.appendChild(capturing)
+        const capturingPromise = new Promise((resolve) => {
+            const capturing = document.createElement('script')
+            capturing.async = true
+            capturing.id = 'progressive-web-capture'
+            capturing.src = CAPTURING_CDN
+            capturing.onload = () => {
+                window.Capture.init((capture) => {
+                    // NOTE: by this time, the captured doc has changed a little bit from original desktop.
+                    // It now has some of our own assets (e.g. main.css) but they can be safely ignored.
+                    window.Progressive.initialCapturedDocHTML = capture.enabledHTMLString()
+                    resolve()
+                })
+            }
+            body.appendChild(capturing)
+        })
+
+        Promise.all([capturingPromise]).then(() => loadMain(body))
     })
 } else {
     const capturing = document.createElement('script')