+++ web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## To be released
+
+## 0.15.0 (March 10, 2017)
+- Fix for add-to-cart when Magento Full Page Cache is enabled [#406](https://github.com/mobify/progressive-web-scaffold/pull/406)
+
+## 0.14.0 (March 3, 2017)
+- Add support for Progressive Mobile Apps
+
 ## 0.13.1 (March 1, 2017)
 - Fix for cart fetch when Magento backend has full page cache enabled